fix type (int -> long) for more Amount fields

+++ src/Telegram.Bot/Requests/Payments/CreateInvoiceLinkRequest.cs
@@ -39,7 +39,7 @@ public partial class CreateInvoiceLinkRequest() : RequestBase<string>("createInv
 
     /// <summary>The maximum accepted amount for tips in the <em>smallest units</em> of the currency (integer, <b>not</b> float/double). For example, for a maximum tip of <c>US$ 1.45</c> pass <c><see cref="MaxTipAmount">MaxTipAmount</see> = 145</c>. See the <em>exp</em> parameter in <a href="https://core.telegram.org/bots/payments/currencies.json">currencies.json</a>, it shows the number of digits past the decimal point for each currency (2 for the majority of currencies). Defaults to 0. Not supported for payments in <a href="https://t.me/BotNews/90">Telegram Stars</a>.</summary>
     [JsonPropertyName("max_tip_amount")]
-    public int? MaxTipAmount { get; set; }
+    public long? MaxTipAmount { get; set; }
 
     /// <summary>A array of suggested amounts of tips in the <em>smallest units</em> of the currency (integer, <b>not</b> float/double). At most 4 suggested tip amounts can be specified. The suggested tip amounts must be positive, passed in a strictly increased order and must not exceed <see cref="MaxTipAmount">MaxTipAmount</see>.</summary>
     [JsonPropertyName("suggested_tip_amounts")]
